@@ -27,18 +27,18 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "~8.2.0",
+    "@angular/animations": "~9.1.0",
     "@angular/cdk": "^8.2.0",
-    "@angular/common": "~8.2.0",
-    "@angular/compiler": "~8.2.0",
-    "@angular/core": "~8.2.0",
-    "@angular/elements": "^8.2.7",
+    "@angular/common": "~9.1.0",
+    "@angular/compiler": "~9.1.0",
+    "@angular/core": "~9.1.0",
+    "@angular/elements": "^9.1.0",
     "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "~8.2.0",
+    "@angular/forms": "~9.1.0",
     "@angular/material": "^8.2.0",
-    "@angular/platform-browser": "~8.2.0",
-    "@angular/platform-browser-dynamic": "~8.2.0",
-    "@angular/router": "~8.2.0",
+    "@angular/platform-browser": "~9.1.0",
+    "@angular/platform-browser-dynamic": "~9.1.0",
+    "@angular/router": "~9.1.0",
     "@aspnet/signalr": "^1.1.4",
     "@lcu/common": "^0.71.335",
     "@lowcodeunit-devkit/ng-lcu": "^0.82.425",
@@ -50,21 +50,21 @@
     "hammerjs": "^2.0.8",
     "ng2-charts": "^2.3.0",
     "ngx-build-plus": "^8.1.4",
-    "rxjs": "~6.4.0",
+    "rxjs": "~6.5.4",
     "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.25",
-    "@angular-devkit/build-ng-packagr": "~0.803.25",
-    "@angular/cli": "~8.3.25",
-    "@angular/compiler-cli": "~8.2.0",
-    "@angular/language-service": "~8.2.0",
+    "@angular-devkit/build-angular": "~0.901.0",
+    "@angular-devkit/build-ng-packagr": "~0.901.0",
+    "@angular/cli": "~9.1.0",
+    "@angular/compiler-cli": "~9.1.0",
+    "@angular/language-service": "~9.1.0",
     "@lcu/cli": "^0.23.8",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "concat-glob-cli": "^0.1.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -74,12 +74,11 @@
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
     "mkdirp": "^0.5.1",
-    "ng-packagr": "^5.3.0",
+    "ng-packagr": "^9.0.0",
     "protractor": "~5.4.0",
     "rimraf": "^3.0.0",
     "ts-node": "~7.0.0",
-    "tsickle": "^0.36.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "typescript": "~3.8.3"
   }
 }
